+++ decidim-admin/app/permissions/decidim/admin/permissions.rb
@@ -44,6 +44,7 @@ module Decidim
           allow! if permission_action.subject == :authorization
           allow! if permission_action.subject == :authorization_workflow
           allow! if permission_action.subject == :static_page_topic
+          allow! if permission_action.subject == :help_sections
         end
 
         permission_action